Screenshot of the project in LensStudio
Version 1.0 of the "interactive grid"

This screenshot shows the SceneObjects and resources that need to be created. Once created, they need to be connected to the script.js attached to the "engine" object.

A set of 'full body colliders' needs to be added to be detected by the balls in the grid.